# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: Jalapão Store API
  version: 1.0.0
paths:
...
  /supplies/:
    get:
      operationId: supplies_list
      description: 'Sem DELETE: insumo sai de uso desativando (`active`).'
      parameters:
      - in: query
        name: active
        schema:
          type: boolean
      - in: query
        name: category
        schema:
          type: string
          format: uuid
      - in: query
        name: in_stock
        schema:
          type: boolean
      - in: query
        name: material
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Qual campo usar ao ordenar os resultados.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Um número de página dentro do conjunto de resultados paginado.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: Um termo de busca.
        schema:
          type: string
      tags:
      - supplies
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSupplyList'
          description: ''
    post:
      operationId: supplies_create
      description: 'Sem DELETE: insumo sai de uso desativando (`active`).'
      tags:
      - supplies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Supply'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Supply'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Supply'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supply'
          description: ''
...
components:
  schemas:
...
    PaginatedSupplyList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Supply'
...
    Supply:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        category:
          type: string
          format: uuid
        category_name:
          type: string
          readOnly: true
        category_is_filament:
          type: boolean
          readOnly: true
        name:
          type: string
          maxLength: 120
        unit:
          type: string
          maxLength: 30
        notes:
          type: string
          maxLength: 500
        active:
          type: boolean
        material:
          type: string
          maxLength: 40
        color:
          type: string
          maxLength: 40
        roll_weight_g:
          type: string
          format: decimal
          pattern: ^-?\d{0,7}(?:\.\d{0,3})?$
          nullable: true
        roll_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          nullable: true
        price_per_gram:
          type: string
          nullable: true
          readOnly: true
        price_per_kg:
          type: string
          nullable: true
          readOnly: true
        quantity:
          type: integer
          readOnly: true
          default: 0
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - category
      - category_is_filament
      - category_name
      - created_at
      - id
      - name
      - price_per_gram
      - price_per_kg
      - quantity
      - updated_at
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: /jalapao-store/backend-api
  description: Produção
- url: /api/v1
  description: Backend direto (desenvolvimento)